PLAYG-16 Refactor Res. Classes (Extract to Files)
